Refactor to use controller.
